Send - added validation for paste address #669
